Refactored_zeyu/requirement.txt
numpy==1.26.4
pandas==2.2.0
PyYAML==6.0.1